+++ src/reset.rs
@@ -59,9 +59,9 @@ impl<T: Reset> Reset for &T {
 impl<T: Reset> Reset for Option<T> {
     #[inline]
     fn system_reset(&self, reset_type: u32, reset_reason: u32) -> SbiRet {
-        self.as_ref()
-            .map(|inner| T::system_reset(inner, reset_type, reset_reason))
-            .unwrap_or(SbiRet::not_supported())
+        self.as_ref().map_or(SbiRet::not_supported(), |inner| {
+            T::system_reset(inner, reset_type, reset_reason)
+        })
     }
     #[inline]
     fn _rustsbi_probe(&self) -> usize {